make it better 3

+++ src/components/ListEmployeeComponent.js
@@ -27,33 +27,42 @@ export default class ListEmployeeComponent extends Component {
     }
 
     deleteEmployee(id) {
-        EmployeeService.deleteEmployee(id).then(res => {
-            this.setState({ employees: this.state.employees.filter(employee => employee.employeeId !== id) });
-        });
+        const r = window.confirm("Do you really want to delete this employee information?"); if (r == true) {
+            EmployeeService.deleteEmployee(id).then(res => {
+                this.setState({ employees: this.state.employees.filter(employee => employee.employeeId !== id) });
+            });
+        }
+
     }
 
-    viewEmployee(id){
+    viewEmployee(id) {
         this.props.history.push(`/view-employee/${id}`);
     }
 
     render() {
         return (
-            <div>
+            <div style={{marginBottom:"50px"}}>
                 <div>
-                    <h2 className="text-center">Employee List</h2>
+                    <h2 className="text-center">Employee Database ( Total : {this.state.employees.length} )</h2>
                 </div>
                 <div>
-                    <button className="btn btn-primary" onClick={this.addEmployee}>Add Employee</button>
+                    <i className="btn btn-success btn-lg fas fa-plus-circle"
+                        onClick={this.addEmployee}>
+                        <span style={{ fontFamily: 'fantasy', marginLeft: "10px" }}>Add Employee</span>
+                    </i>
                 </div>
 
                 <div>
                     <table className="table table-striped">
                         <thead>
                             <tr>
-                                <th scope="col">Id</th>
+                                <th scope="col">Photo </th>
                                 <th scope="col">Name</th>
-                                <th scope="col">Salary</th>
-                                <th scope="col">Update</th>
+                                <th scope="col">Employee Id</th>
+                                <th scope="col">Department</th>
+                                <th scope="col">Joining date</th>
+                                <th scope="col">View/Edit</th>
+                                <th scope="col">Delete</th>
                             </tr>
                         </thead>
                         <tbody>
@@ -61,16 +70,35 @@ export default class ListEmployeeComponent extends Component {
                                 this.state.employees.map(
                                     employee =>
                                         <tr key={employee.employeeId}>
+                                            <td><img src={employee.photoUrl} 
+                                                onClick={() => this.viewEmployee(employee.employeeId)} 
+                                                style={{ width: "50px", height: "50px" , cursor:'pointer'}}></img>
+                                            </td>
+                                            
+                                            <td><b>{employee.employeeName}</b></td>
+                                            
                                             <td>{employee.employeeId}</td>
-                                            <td>{employee.employeeName}</td>
-                                            <td>{employee.salary}</td>
+                                            
+                                            <td>{employee.department}</td>
+                                            
+                                            <td>{employee.dateOfJoining != null ? employee.dateOfJoining.toString().split('T')[0] : " - "}</td>
+                                            
                                             <td>
-                                                <button className="btn btn-dark"
-                                                    onClick={() => this.updateEmployee(employee.employeeId)}>Update</button>
-                                                <button className="btn btn-danger" style={{ marginLeft: "10px" }}
-                                                    onClick={() => this.deleteEmployee(employee.employeeId)}>Delete</button>
-                                                <button className="btn btn-info" style={{ marginLeft: "10px" }}
-                                                    onClick={() => this.viewEmployee(employee.employeeId)}>View</button>
+                                                <button className="btn"
+                                                    onClick={() => this.viewEmployee(employee.employeeId)}>
+                                                    <i className="fa fa-info-circle" aria-hidden="true"></i>
+                                                </button>
+
+                                                <button className="btn"
+                                                    onClick={() => this.updateEmployee(employee.employeeId)}>
+                                                    <i className="fa fa-edit" aria-hidden="true"></i>
+                                                </button>
+                                            </td>
+                                            <td>
+                                                <button className="btn btn-danger" style={{ marginLeft: "0px" }}
+                                                    onClick={() => this.deleteEmployee(employee.employeeId)}>
+                                                    <i className="fa fa-trash" aria-hidden="true"></i>
+                                                </button>
                                             </td>
                                         </tr>
                                 )
@@ -79,7 +107,8 @@ export default class ListEmployeeComponent extends Component {
                     </table>
                 </div>
 
-            </div>
+
+            </div >
         )
     }
 }
